ms-multi-data-handeling
This is a program a friend and I wrote to go into each subfolder of a directory and plot every .csv data there. They were MS-Data, so I wrote a script to set a threshold and label only peaks that are a bit further apart. Use, or comment if you like. I am not a coder; I have no idea what I am doing. I just keep googling, and it works.